moved tests to separate package
