+++ src/test/java/ru/yandex/practicum/stellarburger/BaseTest.java
@@ -5,7 +5,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
 
 public class BaseTest {
-    ChromeDriver driver;
+    protected ChromeDriver driver;
 
     public void setUpBrowser() {
         /*ChromeOptions options = new ChromeOptions();
