+++ src/test/java/ru/yandex/practicum/stellarburger/tests/RegisterUserTest.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.stellarburger;
+package ru.yandex.practicum.stellarburger.tests;
 
 
 import io.qameta.allure.Feature;
@@ -6,6 +6,7 @@ import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import ru.yandex.practicum.stellarburger.BaseTest;
 import ru.yandex.practicum.stellarburger.api.User;
 import ru.yandex.practicum.stellarburger.api.UserClient;
 import ru.yandex.practicum.stellarburger.api.UserCredentials;
